[data] Resample to seconds before aggregating
The problem is that for stat fcts like "mean",
pandas doesn't actually compute the time average, but only averages the existing
values in the time window. This doesn't make
sense for Home Assistant sensors
that only report state when it changes.
To fix this, I now resample to 1 second and padding to the right before
aggregating.